+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Main
+* [5848](https://github.com/grafana/loki/pull/5848) **arcosx** add Baidu AI Cloud as storage backends for Loki
 * [5799](https://github.com/grafana/loki/pull/5799) **cyriltovena** Fix deduping issues when multiple entries with the same timestamp exist.
 * [5799](https://github.com/grafana/loki/pull/5799) **cyriltovena** Fixes deduping issues when multiple entries exists with the same timestamp.
 * [5780](https://github.com/grafana/loki/pull/5780) **simonswine**: Update alpine image to 3.15.4.